Split metadata.rs into mod.rs and value.rs
